+++ spec/puppet-check/puppet_parser_spec.rb
@@ -23,14 +23,14 @@ describe PuppetParser do
       PuppetCheck.style_check = true
       PuppetParser.manifest([fixtures_dir + 'manifests/style_parser.pp'])
       expect(PuppetCheck.error_files).to eql([])
-      expect(PuppetCheck.warning_files[0]).to match(%r{^\-\- #{fixtures_dir}manifests/style_parser.pp:\nUnrecognized escape sequence.*\nUnrecognized escape sequence.*\ndouble quoted string containing})
+      expect(PuppetCheck.warning_files[0]).to match(%r{^\-\- #{fixtures_dir}manifests/style_parser.pp:\nUnrecognized escape sequence.*\nUnrecognized escape sequence.*\n.*double quoted string containing})
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad lint style Puppet manifest in the warning files array' do
       PuppetCheck.style_check = true
       PuppetParser.manifest([fixtures_dir + 'manifests/style_lint.pp'])
       expect(PuppetCheck.error_files).to eql([])
-      expect(PuppetCheck.warning_files[0]).to match(%r{^\-\- #{fixtures_dir}manifests/style_lint.pp:\ndouble quoted string containing.*\nindentation of})
+      expect(PuppetCheck.warning_files[0]).to match(%r{^\-\- #{fixtures_dir}manifests/style_lint.pp:\n.*double quoted string containing.*\n.*indentation of})
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style Puppet manifest in the clean files array when puppetlint_args ignores its warnings' do